댓글, 좋아요 동일유저판별오류 수정 #6

+++ src/main/java/com/innovation/stockstock/controller/CommentController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class CommentController {
 
-    private final CommentService commentService; // 왜 final을 붙여야 인식이 되는걸까?
+    private final CommentService commentService; // final
 
     @PostMapping(value = "/api/member/comment")
     public ResponseEntity<Object> postComment(@RequestBody CommentDto commentDto) {
@@ -18,12 +18,12 @@ public class CommentController {
     }
 
     @PutMapping(value = "/api/member/comment/{commentId}")
-    public ResponseEntity<Object> putComment(@PathVariable Long commentId, @RequestBody CommentDto commentDto) {
+    public ResponseEntity<Object> putComment(@PathVariable int commentId, @RequestBody CommentDto commentDto) {
         return commentService.putComment(commentId,commentDto);
     }
 
     @DeleteMapping(value = "/api/member/comment/{commentId}")
-    public ResponseEntity<Object> deleteComment(@PathVariable Long commentId) {
+    public ResponseEntity<Object> deleteComment(@PathVariable int commentId) {
         return commentService.deleteComment(commentId);
     }
