+++ src/main/java/com/innovation/stockstock/dto/CommentDto.java
@@ -5,6 +5,6 @@ import lombok.Getter;
 @Getter
 public class CommentDto {
 
-    private Long postId;
+    private int postId;
     private String content;
 }
